Restructured build
Each servlet implementation now has its own project for isolation.
