+++ reactive-streams-servlet-test-jetty/build.gradle
@@ -1,0 +1,13 @@
+description = "Jetty verification for the Servlet API implementation of Reactive Streams"
+version = "1.0.0"
+
+dependencies {
+  testCompile project(":reactive-streams-servlet-test")
+  testCompile "org.eclipse.jetty:jetty-server:$jettyVersion"
+}
+
+test {
+  // Disabled due to https://github.com/eclipse/jetty.project/issues/2006
+  exclude '**/JettyRequestPublisherTest.class'
+}
+
