+++ reactive-streams-servlet-test-jetty/src/test/java/org/reactivestreams/servlet/test/jetty/WithJettyServer.java
@@ -1,9 +1,20 @@
-package org.reactivestreams.servlet;
+/************************************************************************
+ * Licensed under Public Domain (CC0)                                    *
+ *                                                                       *
+ * To the extent possible under law, the person who associated CC0 with  *
+ * this code has waived all copyright and related or neighboring         *
+ * rights to this code.                                                  *
+ *                                                                       *
+ * You should have received a copy of the CC0 legalcode along with this  *
+ * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.*
+ ************************************************************************/
+package org.reactivestreams.servlet.test.jetty;
 
 import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
 import org.eclipse.jetty.server.handler.AbstractHandler;
+import org.reactivestreams.servlet.test.WithVerificationServer;
 import org.testng.annotations.AfterClass;
 
 import javax.servlet.ServletException;
