+++ reactive-streams-servlet-test-tomcat/src/test/java/org/reactivestreams/servlet/test/tomcat/TomcatResponseSubscriberBlackboxTest.java
@@ -1,0 +1,17 @@
+/************************************************************************
+ * Licensed under Public Domain (CC0)                                    *
+ *                                                                       *
+ * To the extent possible under law, the person who associated CC0 with  *
+ * this code has waived all copyright and related or neighboring         *
+ * rights to this code.                                                  *
+ *                                                                       *
+ * You should have received a copy of the CC0 legalcode along with this  *
+ * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.*
+ ************************************************************************/
+package org.reactivestreams.servlet.test.tomcat;
+
+import org.reactivestreams.servlet.test.AbstractResponseSubscriberBlackboxTest;
+import org.testng.annotations.Test;
+
+@Test
+public class TomcatResponseSubscriberBlackboxTest extends AbstractResponseSubscriberBlackboxTest implements WithTomcatServer { }
