+++ reactive-streams-servlet-test/src/main/java/org/reactivestreams/servlet/test/AbstractRequestPublisherTest.java
@@ -8,11 +8,12 @@
  * You should have received a copy of the CC0 legalcode along with this  *
  * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.*
  ************************************************************************/
-package org.reactivestreams.servlet;
+package org.reactivestreams.servlet.test;
 
 import org.eclipse.jetty.client.HttpClient;
 import org.eclipse.jetty.client.api.ContentProvider;
 import org.reactivestreams.Publisher;
+import org.reactivestreams.servlet.RequestPublisher;
 import org.reactivestreams.tck.PublisherVerification;
 import org.testng.annotations.*;
