+++ reactive-streams-servlet-test/src/main/java/org/reactivestreams/servlet/test/Probes.java
@@ -8,7 +8,7 @@
  * You should have received a copy of the CC0 legalcode along with this  *
  * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.*
  ************************************************************************/
-package org.reactivestreams.servlet;
+package org.reactivestreams.servlet.test;
 
 import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
