+++ reactive-streams-servlet-test/src/main/java/org/reactivestreams/servlet/test/ServletTestEnvironment.java
@@ -1,0 +1,17 @@
+/************************************************************************
+ * Licensed under Public Domain (CC0)                                    *
+ *                                                                       *
+ * To the extent possible under law, the person who associated CC0 with  *
+ * this code has waived all copyright and related or neighboring         *
+ * rights to this code.                                                  *
+ *                                                                       *
+ * You should have received a copy of the CC0 legalcode along with this  *
+ * work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.*
+ ************************************************************************/
+package org.reactivestreams.servlet.test;
+
+import org.reactivestreams.tck.TestEnvironment;
+
+public class ServletTestEnvironment {
+  public static final TestEnvironment INSTANCE = new TestEnvironment(100);
+}
