--- src/test/java/org/reactivestreams/servlet/JettyRequestPublisherTest.java
@@ -1,6 +1,0 @@
-package org.reactivestreams.servlet;
-
-import org.testng.annotations.Test;
-
-@Test
-public class JettyRequestPublisherTest extends AbstractRequestPublisherTest implements WithJettyServer { }
