--- src/test/java/org/reactivestreams/servlet/JettyResponseSubscriberBlackboxTest.java
@@ -1,6 +1,0 @@
-package org.reactivestreams.servlet;
-
-import org.testng.annotations.Test;
-
-@Test
-public class JettyResponseSubscriberBlackboxTest extends AbstractResponseSubscriberBlackboxTest implements WithJettyServer { }
