--- src/test/java/org/reactivestreams/servlet/JettyResponseSubscriberWhiteboxTest.java
@@ -1,6 +1,0 @@
-package org.reactivestreams.servlet;
-
-import org.testng.annotations.Test;
-
-@Test
-public class JettyResponseSubscriberWhiteboxTest extends AbstractResponseSubscriberWhiteboxTest implements WithJettyServer { }
