--- src/test/java/org/reactivestreams/servlet/TomcatRequestPublisherTest.java
@@ -1,6 +1,0 @@
-package org.reactivestreams.servlet;
-
-import org.testng.annotations.Test;
-
-@Test
-public class TomcatRequestPublisherTest extends AbstractRequestPublisherTest implements WithTomcatServer { }
