--- src/test/java/org/reactivestreams/servlet/TomcatResponseSubscriberBlackboxTest.java
@@ -1,6 +1,0 @@
-package org.reactivestreams.servlet;
-
-import org.testng.annotations.Test;
-
-@Test
-public class TomcatResponseSubscriberBlackboxTest extends AbstractResponseSubscriberBlackboxTest implements WithTomcatServer { }
