--- src/test/java/org/reactivestreams/servlet/TomcatResponseSubscriberWhiteboxTest.java
@@ -1,6 +1,0 @@
-package org.reactivestreams.servlet;
-
-import org.testng.annotations.Test;
-
-@Test
-public class TomcatResponseSubscriberWhiteboxTest extends AbstractResponseSubscriberWhiteboxTest implements WithTomcatServer { }
